fix: use yarn to install deps when linking yarn plugins
