+++ package.json
@@ -12,7 +12,8 @@
     "npm-package-arg": "^11.0.1",
     "npm-run-path": "^5.3.0",
     "semver": "^7.6.0",
-    "validate-npm-package-name": "^5.0.0"
+    "validate-npm-package-name": "^5.0.0",
+    "yarn": "^1.22.22"
   },
   "devDependencies": {
     "@commitlint/config-conventional": "^18",
