+++ src/npm.ts
@@ -1,101 +1,16 @@
-import {Errors, Interfaces, ux} from '@oclif/core'
+import {Interfaces, ux} from '@oclif/core'
 import makeDebug from 'debug'
-import {fork as cpFork} from 'node:child_process'
 import {readFile} from 'node:fs/promises'
 import {createRequire} from 'node:module'
 import {join, sep} from 'node:path'
-import {npmRunPathEnv} from 'npm-run-path'
 
+import {ExecOptions, Output, fork} from './fork.js'
 import {LogLevel} from './log-level.js'
 
 const debug = makeDebug('@oclif/plugin-plugins:npm')
 
-type ExecOptions = {
-  cwd: string
-  logLevel: LogLevel
-}
-
 type InstallOptions = ExecOptions & {
   prod?: boolean
-}
-
-export type NpmOutput = {
-  stderr: string[]
-  stdout: string[]
-}
-
-async function fork(modulePath: string, args: string[] = [], {cwd, logLevel}: ExecOptions): Promise<NpmOutput> {
-  return new Promise((resolve, reject) => {
-    const forked = cpFork(modulePath, args, {
-      cwd,
-      env: {
-        ...npmRunPathEnv(),
-        // Disable husky hooks because a plugin might be trying to install them, which will
-        // break the install since the install location isn't a .git directory.
-        HUSKY: '0',
-      },
-      execArgv: process.execArgv
-        .join(' ')
-        // Remove --loader ts-node/esm from execArgv so that the subprocess doesn't fail if it can't find ts-node.
-        // The ts-node/esm loader isn't need to execute npm commands anyways.
-        .replace('--loader ts-node/esm', '')
-        .replace('--loader=ts-node/esm', '')
-        .split(' ')
-        .filter(Boolean),
-      stdio: [0, null, null, 'ipc'],
-    })
-
-    const possibleLastLinesOfNpmInstall = ['up to date', 'added']
-    const stderr: string[] = []
-    const stdout: string[] = []
-    const loggedStderr: string[] = []
-    const loggedStdout: string[] = []
-
-    const shouldPrint = (str: string): boolean => {
-      // For ux cleanliness purposes, don't print the final line of npm install output if
-      // the log level is 'notice' and there's no other output.
-      const noOtherOutput = loggedStderr.length === 0 && loggedStdout.length === 0
-      const isLastLine = possibleLastLinesOfNpmInstall.some((line) => str.startsWith(line))
-      if (noOtherOutput && isLastLine && logLevel === 'notice') {
-        return false
-      }
-
-      return logLevel !== 'silent'
-    }
-
-    forked.stderr?.setEncoding('utf8')
-    forked.stderr?.on('data', (d: Buffer) => {
-      const output = d.toString().trim()
-      stderr.push(output)
-      if (shouldPrint(output)) {
-        loggedStderr.push(output)
-        ux.log(output)
-      } else debug(output)
-    })
-
-    forked.stdout?.setEncoding('utf8')
-    forked.stdout?.on('data', (d: Buffer) => {
-      const output = d.toString().trim()
-      stdout.push(output)
-      if (shouldPrint(output)) {
-        loggedStdout.push(output)
-        ux.log(output)
-      } else debug(output)
-    })
-
-    forked.on('error', reject)
-    forked.on('exit', (code: number) => {
-      if (code === 0) {
-        resolve({stderr, stdout})
-      } else {
-        reject(
-          new Errors.CLIError(`${modulePath} ${args.join(' ')} exited with code ${code}`, {
-            suggestions: ['Run with DEBUG=@oclif/plugin-plugins* to see debug output.'],
-          }),
-        )
-      }
-    })
-  })
 }
 
 export class NPM {
@@ -108,7 +23,7 @@ export class NPM {
     this.logLevel = logLevel
   }
 
-  async exec(args: string[] = [], options: ExecOptions): Promise<NpmOutput> {
+  async exec(args: string[] = [], options: ExecOptions): Promise<Output> {
     const bin = await this.findNpm()
     debug('npm binary path', bin)
 
@@ -130,20 +45,20 @@ export class NPM {
     }
   }
 
-  async install(args: string[], opts: InstallOptions): Promise<NpmOutput> {
+  async install(args: string[], opts: InstallOptions): Promise<Output> {
     const prod = opts.prod ? ['--omit', 'dev'] : []
     return this.exec(['install', ...args, ...prod, '--no-audit'], opts)
   }
 
-  async uninstall(args: string[], opts: ExecOptions): Promise<NpmOutput> {
+  async uninstall(args: string[], opts: ExecOptions): Promise<Output> {
     return this.exec(['uninstall', ...args], opts)
   }
 
-  async update(args: string[], opts: ExecOptions): Promise<NpmOutput> {
+  async update(args: string[], opts: ExecOptions): Promise<Output> {
     return this.exec(['update', ...args], opts)
   }
 
-  async view(args: string[], opts: ExecOptions): Promise<NpmOutput> {
+  async view(args: string[], opts: ExecOptions): Promise<Output> {
     return this.exec(['view', ...args], {...opts, logLevel: 'silent'})
   }
